// src/api/generated/model/feedbackCreatedEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackCreatedEventType } from './feedbackCreatedEventType';
import type { FeedbackCreatedPayload } from './feedbackCreatedPayload';

export interface FeedbackCreatedEvent {
  id: string;
  type: FeedbackCreatedEventType;
  payload: FeedbackCreatedPayload;
  created_at: string;
  processed_at: string | null;
}