// src/api/generated/model/feedbackCreatedPayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackCreatedPayloadTargetType } from './feedbackCreatedPayloadTargetType';
import type { FeedbackCreatedPayloadVerdict } from './feedbackCreatedPayloadVerdict';

export interface FeedbackCreatedPayload {
  feedback_id: string;
  user_id: string;
  target_type: FeedbackCreatedPayloadTargetType;
  target_id: string;
  service_id: string;
  verdict: typeof FeedbackCreatedPayloadVerdict[keyof typeof FeedbackCreatedPayloadVerdict];
  reason: string | null;
}