// src/api/generated/model/runFinishedEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { RunFinishedEventType } from './runFinishedEventType';
import type { RunFinishedPayload } from './runFinishedPayload';

export interface RunFinishedEvent {
  id: string;
  type: RunFinishedEventType;
  payload: RunFinishedPayload;
  created_at: string;
  processed_at: string | null;
}